backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
certifi==2025.7.9
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
deprecation==2.1.0
fastapi==0.116.0
grpcio==1.73.1
grpcio-health-checking==1.73.1
grpcio-tools==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==6.0.0
packaging==25.0
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-docx==1.2.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
validators==0.34.0
watchdog==6.0.0
weaviate-client==4.15.4
langchain
numpy
pandas
tiktoken
